agent/requirements.txt
psycopg2-binary
ping3
requests
